schema: move schema_extension into schema_fwd.hh
schema_extension is a small (~18-line) pure-virtual interface, but
living in schema.hh (1115 lines) meant every consumer that only needs
to subclass it - tombstone_gc_extension.hh, cdc/cdc_extension.hh,
db/paxos_grace_seconds_extension.hh, db/per_partition_rate_limit_extension.hh,
db/tags/extension.hh - was pulling in the full schema.hh (421 direct
includers project-wide) just for this one class.

Move schema_extension into schema_fwd.hh (which schema.hh already
includes, so nothing loses visibility to it) and repoint the five
extension headers at schema_fwd.hh, adding the couple of small
includes (bytes_fwd.hh, <map>, <string>) they were previously getting
transitively through schema.hh.

Targets full-build time via reduced transitive fan-out from schema.hh.
