+++ schema/schema_fwd.hh
@@ -8,8 +8,13 @@
 
 #pragma once
 
-#include <seastar/core/shared_ptr.hh>
+#include <map>
+#include <string>
 
+#include <seastar/core/shared_ptr.hh>
+#include <seastar/core/future.hh>
+
+#include "bytes_fwd.hh"
 #include "utils/UUID.hh"
 #include "utils/UUID_gen.hh"
 
@@ -19,9 +24,44 @@ using column_count_type = uint32_t;
 using column_id = column_count_type;
 
 class schema;
-class schema_extension;
 
 using schema_ptr = seastar::lw_shared_ptr<const schema>;
+
+/**
+ * Schema extension. An opaque type representing
+ * entries in the "extensions" part of a table/view (see schema_tables).
+ *
+ * An extension has a name (the mapping key), and it can re-serialize
+ * itself to bytes again, when we write back into schema tables.
+ *
+ * Code using a particular extension can locate it by name in the schema map,
+ * and barring the "is_placeholder" says true, cast it to whatever might
+ * be the expected implementation.
+ *
+ * We allow placeholder object since an extension written to schema tables
+ * might be unavailable on next boot/other node. To avoid losing the config data,
+ * a placeholder object is put into schema map, which at least can
+ * re-serialize the data back.
+ *
+ */
+class schema_extension {
+public:
+    virtual ~schema_extension() {};
+    [[deprecated("Use dedicated columns in system_schema.scylla_tables instead")]]
+    schema_extension() = default;
+
+    virtual seastar::future<> validate(const schema&) const {
+        return seastar::make_ready_future<>();
+    }
+    virtual bytes serialize() const = 0;
+    virtual bool is_placeholder() const {
+        return false;
+    }
+    using default_map_type = std::map<seastar::sstring, seastar::sstring>;
+    // default impl assumes options are in a map.
+    // implementations should override if not
+    virtual std::string options_to_string() const;
+};
 
 using table_id = utils::tagged_uuid<struct table_id_tag>;
